+++ src/renderer/ScreenController.tsx
@@ -21,6 +21,7 @@ import { BackendRemote } from './backend-com'
 import { debouncedUpdateBadgeCounter } from './system-integration/badge-counter'
 import { hintUpdateIfNessesary, updateDeviceChats } from './deviceMessages'
 import { runtime } from './runtime'
+import { DcEventType } from '@deltachat/jsonrpc-client'
 
 const log = getLogger('renderer/ScreenController')
 
@@ -148,9 +149,7 @@ export default class ScreenController extends Component {
   }
 
   componentDidMount() {
-    ipcRenderer.on('error', this.onError)
-    ipcRenderer.on('DC_EVENT_ERROR', this.onError)
-    ipcRenderer.on('success', this.onSuccess)
+    BackendRemote.on('Error', this.onError)
     ipcRenderer.on('showAboutDialog', this.onShowAbout)
     ipcRenderer.on('showKeybindingsDialog', this.onShowKeybindings)
     ipcRenderer.on('showSettingsDialog', this.onShowSettings)
@@ -162,19 +161,20 @@ export default class ScreenController extends Component {
   }
 
   componentWillUnmount() {
+    BackendRemote.off('Error', this.onError)
     ipcRenderer.removeListener('showAboutDialog', this.onShowAbout)
     ipcRenderer.removeListener('showSettingsDialog', this.onShowSettings)
-    ipcRenderer.removeListener('error', this.onError)
-    ipcRenderer.removeListener('DC_EVENT_ERROR', this.onError)
-    ipcRenderer.removeListener('success', this.onSuccess)
     ipcRenderer.removeListener('open-url', this.onOpenUrl)
   }
 
-  onError(_event: any, [data1, data2]: [string | number, string]) {
-    if (this.state.screen === Screens.Welcome) return
-    if (data1 === 0) data1 = ''
-    const text = data1 + data2
-    this.userFeedback({ type: 'error', text })
+  onError(accountId: number, { msg }: DcEventType<'Error'>) {
+    if (
+      this.selectedAccountId !== accountId ||
+      this.state.screen === Screens.Welcome
+    ) {
+      return
+    }
+    this.userFeedback({ type: 'error', text: msg })
   }
 
   onSuccess(_event: any, text: string) {