+++ src/shared/logger.ts
@@ -47,6 +47,20 @@ export function printProcessLogLevelInfo() {
     )}`,
     emojiFontCss
   )
+  /* ignore-console-log */
+  console.info(
+    `# Tips and Tricks for working with the browser console:
+## Use the search to filter the output like:
+space seperate terms, exclude with -, if your term contains spaces you should exape it with "
+
+-👻                 don't show events from background accounts (not selected accounts)
+-📡                 don't show any events
+-renderer/jsonrpc   don't show jsonrpc messages
+renderer/jsonrpc    show only jsonrpc messages
+
+Start deltachat with --devmode (or --log-debug and --log-to-console) argument to show full log output.
+  `
+  )
 }
 
 export type LogHandlerFunction = (